+++ src/utils/add-modulo.js
@@ -1,0 +1,6 @@
+export function addModulo(n, modulo) {
+  return (n + 1) % modulo;
+}
+export function subModulo(n, modulo) {
+  return (n - 1 + modulo) % modulo;
+}